fstab: do not check f2fs
It is not needed to check f2fs before mounting the filesystem.
The kernel include a not up to date f2fs and fsck seems to possibly
create more damages regarding users feedback.

fsck was not used/available on Android 5.1 so this is not a real issue.

This speed up the boot process (time on Sony logo)